rename controller soldier
