+++ api/controller/magazine_controller.go
@@ -1,0 +1,1 @@
+package controller
