+++ api/service/magazine_service.go
@@ -1,0 +1,1 @@
+package service
